test(hyperdim): add AVX2 hamming edge case tests, benchmark, extract to fix LOC gate
- Extract 21-unit-test #[cfg(test)] mod from hyperdim_simd.rs (613→317 LOC)
  into hyperdim_simd_tests.rs via #[path] attribute — under 500 LOC gate
- Add 7 new AVX2 hamming distance edge case tests: all_ones_vs_zeros,
  identical, complements, single_bit, alternating_pattern, random_100
  (expanded from 10), sparse_pattern, mixed_popcounts — each validates
  against scalar hamming_distance_optimized for correctness
- Add bench_hamming_distance benchmark group (public API path) in
  benches/benchmark.rs — validates runtime SIMD dispatch performance
- Addresses codacy review comments on PR #262: correctness tests for
  various vector sizes/edge cases + performance benchmarking
